Port ioctl recording to go through the main process
Note that this changes the ioctl_tree_insert API to be more vala
friendly. i.e. it will return the new tree pointer, and it will always
take the passed tree node, free'ing it if a duplicate has been found.

This fixes storing the record if the process overrides SIGINT for
ioctl's. See #96