Implement pcap based USB replay
The ioctl-tree based replay is mostly useful for protocols that are
fundamentally state-less. With state-full protocols, the ioctl-tree may
randomly jump around, cause confusing situations and gets the order
wrong during recording.

The pcap based replay is also able to play-back control transfers adding
this feature to umockdev, even if outside of the original ioctl-tree
implementation.

Note that umockdev does not support creating a recording. Recordings
should be done using a separate tool such as wireshark.

Closes: #82, #117