umockdev-ioctl: Allow overriding read/write methods for devices
This permits overriding the read/write handler for devices. Doing so
permits us to e.g. correctly emulate SPI devices which may be accessed
both through read/write and ioctl.

The default implementation is to fall back, which may e.g. pick the
underlying PTY that was created.

Note that read/write emulation using this method is not done for the
fallback ioctl emulation socket as that would result in uneccessary
overhead for script handlers.